# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Taja Multilingual Support Bot",
    "description": "Python-first multilingual FAQ assistant for Taja customer support.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/languages": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Languages",
        "operationId": "get_languages_api_v1_languages_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LanguageOption"
                  },
                  "type": "array",
                  "title": "Response Get Languages Api V1 Languages Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/faqs": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Get Faqs",
        "operationId": "get_faqs_api_v1_faqs_get",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/LanguageCode",
              "default": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicFaq"
                  },
                  "title": "Response Get Faqs Api V1 Faqs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat",
        "operationId": "chat_api_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/webhooks/gupshup": {
      "post": {
        "tags": [
          "webhooks"
        ],
        "summary": "Gupshup Webhook",
        "operationId": "gupshup_webhook_webhooks_gupshup_post",
        "parameters": [
          {
            "name": "x-webhook-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Webhook-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GupshupWebhookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChatRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Session Id"
          },
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message"
          },
          "language": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LanguageCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "channel": {
            "type": "string",
            "maxLength": 64,
            "title": "Channel",
            "default": "api"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "message"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "language": {
            "$ref": "#/components/schemas/LanguageCode"
          },
          "reply": {
            "type": "string",
            "title": "Reply"
          },
          "intent": {
            "$ref": "#/components/schemas/Intent"
          },
          "matched_faq_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Faq Id"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "requires_human": {
            "type": "boolean",
            "title": "Requires Human",
            "default": false
          },
          "escalation_priority": {
            "$ref": "#/components/schemas/EscalationPriority",
            "default": "none"
          },
          "case_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Reference"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceReference"
            },
            "type": "array",
            "title": "Sources"
          },
          "answer_mode": {
            "type": "string",
            "title": "Answer Mode",
            "default": "deterministic"
          },
          "generation_fallback": {
            "type": "boolean",
            "title": "Generation Fallback",
            "default": false
          },
          "redacted_input": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Redacted Input"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "language",
          "reply",
          "intent",
          "confidence"
        ],
        "title": "ChatResponse"
      },
      "EscalationPriority": {
        "type": "string",
        "enum": [
          "none",
          "normal",
          "high",
          "critical"
        ],
        "title": "EscalationPriority"
      },
      "GupshupDelivery": {
        "properties": {
          "attempted": {
            "type": "boolean",
            "title": "Attempted"
          },
          "delivered": {
            "type": "boolean",
            "title": "Delivered"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "provider_response": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Response"
          }
        },
        "type": "object",
        "required": [
          "attempted",
          "delivered",
          "mode"
        ],
        "title": "GupshupDelivery"
      },
      "GupshupWebhookResponse": {
        "properties": {
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          },
          "chatbot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "delivery": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GupshupDelivery"
              },
              {
                "type": "null"
              }
            ]
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "accepted"
        ],
        "title": "GupshupWebhookResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "app": {
            "type": "string",
            "title": "App"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "knowledge_base_entries": {
            "type": "integer",
            "title": "Knowledge Base Entries"
          }
        },
        "type": "object",
        "required": [
          "status",
          "app",
          "version",
          "knowledge_base_entries"
        ],
        "title": "HealthResponse"
      },
      "Intent": {
        "type": "string",
        "enum": [
          "routine",
          "complaint",
          "fraud_security",
          "account_access",
          "restricted_advice",
          "unsupported",
          "menu"
        ],
        "title": "Intent"
      },
      "LanguageCode": {
        "type": "string",
        "enum": [
          "en",
          "pcm",
          "ha",
          "ig",
          "yo"
        ],
        "title": "LanguageCode"
      },
      "LanguageOption": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/LanguageCode"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "native_name": {
            "type": "string",
            "title": "Native Name"
          },
          "menu_number": {
            "type": "integer",
            "title": "Menu Number"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "native_name",
          "menu_number"
        ],
        "title": "LanguageOption"
      },
      "PublicFaq": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceReference"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "question",
          "answer",
          "sources"
        ],
        "title": "PublicFaq"
      },
      "SourceReference": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "last_verified": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified"
          }
        },
        "type": "object",
        "required": [
          "title",
          "url"
        ],
        "title": "SourceReference"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}